refactor: stop using tsconfig-paths package
It is better to let Vite handle all of the resolving, so this plugin is only responsible for the path mapping.

At the same time, I am also adding the `tsconfck` package, which handles the loading of tsconfig.json files for us. This package is also used in Vite core (we should have Vite expose it in the future, to avoid the duplication).

The `tsconfig-paths` package was previously responsible for config loading and much of the path resolving logic, but now we can ditch it! I'm also ditching the `recrawl-sync` package, which was used for crawling the filesystem in search of tsconfig.json files.

Fixes #54
Closes #73
Probably fixes #63
Probably fixes #68