feat: add `projectDiscovery` option
When this option is set to "lazy", tsconfig files will be loaded on-demand. When an import is resolved, the parent directories of the importer are searched for tsconfig files matching the `configNames` array option (which defaults to ["tsconfig.json", "jsconfig.json"]).

If using TypeScript’s project references, you can ensure they are discovered by adding the main project to the `projects` array option of this plugin. Though, this is only necessary if you’re referencing tsconfig files with non-standard names (e.g. "tsconfig.src.json"). Setting the `projects` option will *not* interfere with lazy project discovery.